# sakina-docs/MODULE_CONTRACTS_OPENAPI_2026-05-28.yaml
openapi: 3.0.3
info:
  title: Sakina Module Contract Shell API
  version: 2026-05-28
  description: >
    Read-only module contract shell endpoints for Quran, Prayer, Knowledge, and Community.
    These are not real content modules yet and must remain gated until verified ingestion and review are implemented.
servers:
  - url: https://api.7jzi.com/v1
paths:
  /modules/quran/overview:
    get:
      summary: Quran module read-only overview shell
      responses:
        "200":
          description: Read-only contract shell
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuranOverview"
        "403":
          $ref: "#/components/responses/FeatureDisabledError"
        "402":
          $ref: "#/components/responses/SubscriptionRequiredError"
  /modules/prayer/overview:
    get:
      summary: Prayer module read-only overview shell
      responses:
        "200":
          description: Read-only contract shell
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrayerOverview"
        "403":
          $ref: "#/components/responses/FeatureDisabledError"
        "402":
          $ref: "#/components/responses/SubscriptionRequiredError"
  /modules/knowledge/overview:
    get:
      summary: Knowledge module read-only overview shell
      responses:
        "200":
          description: Read-only contract shell
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeOverview"
        "403":
          $ref: "#/components/responses/FeatureDisabledError"
        "402":
          $ref: "#/components/responses/SubscriptionRequiredError"
  /modules/community/overview:
    get:
      summary: Community module read-only overview shell
      responses:
        "200":
          description: Read-only contract shell
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityOverview"
        "403":
          $ref: "#/components/responses/FeatureDisabledError"
        "402":
          $ref: "#/components/responses/SubscriptionRequiredError"
  /rag/status:
    get:
      summary: RAG infrastructure status
      responses:
        "200":
          description: RAG status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RagStatusResponse"
  /rag/sources:
    get:
      summary: Verified RAG sources for module
      parameters:
        - in: header
          name: x-sakina-module
          schema:
            type: string
            enum: [quran, prayer, knowledge, community]
      responses:
        "200":
          description: Verified source list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RagSourcesResponse"
        "403":
          $ref: "#/components/responses/FeatureDisabledError"
        "402":
          $ref: "#/components/responses/SubscriptionRequiredError"
  /rag/search:
    get:
      summary: Verified read-only RAG search
      parameters:
        - in: query
          name: module
          required: true
          schema:
            type: string
            enum: [quran, prayer, knowledge, community]
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verified search result items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RagSearchResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "403":
          $ref: "#/components/responses/FeatureDisabledError"
        "402":
          $ref: "#/components/responses/SubscriptionRequiredError"
  /rag/decide:
    post:
      summary: Deterministic Sakina decision algorithm (phase-2 shell)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionRequest"
      responses:
        "200":
          description: Safe decision response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionResponse"
  /iman-journey/{user_id}:
    get:
      summary: Read daily iman journey
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
        - in: query
          name: date
          required: false
          schema:
            type: string
            format: date
        - $ref: "#/components/parameters/UserScopeHeader"
      responses:
        "200":
          description: Daily iman journey
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImanJourneyResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    put:
      summary: Upsert daily iman journey
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
        - $ref: "#/components/parameters/UserScopeHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertImanJourneyRequest"
      responses:
        "200":
          description: Upserted daily iman journey
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImanJourneyResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /iman-journey/{user_id}/privacy:
    get:
      summary: Read iman journey privacy settings
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
        - $ref: "#/components/parameters/UserScopeHeader"
      responses:
        "200":
          description: Privacy settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImanJourneyPrivacySettings"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    put:
      summary: Update iman journey privacy settings
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
        - $ref: "#/components/parameters/UserScopeHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImanJourneyPrivacySettings"
      responses:
        "200":
          description: Updated privacy settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImanJourneyPrivacySettings"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /dua-list/{user_id}:
    get:
      summary: List personal dua items
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
        - $ref: "#/components/parameters/UserScopeHeader"
      responses:
        "200":
          description: Dua list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DuaListResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      summary: Add personal dua item
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
        - $ref: "#/components/parameters/UserScopeHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddDuaItemRequest"
      responses:
        "201":
          description: Created dua item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DuaItem"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
components:
  parameters:
    UserIdPath:
      in: path
      name: user_id
      required: true
      schema:
        type: string
        format: uuid
    UserScopeHeader:
      in: header
      name: x-sakina-user-id
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    ModuleSafetyStatus:
      type: string
      enum:
        - disabled
        - gated
        - enabled_read_only
        - requires_review
    SourceProvenance:
      type: object
      required:
        - source_type
        - source_name
        - source_reference
        - review_status
      properties:
        source_type:
          type: string
        source_name:
          type: string
        source_reference:
          type: string
        review_status:
          type: string
        effective_date:
          type: string
          nullable: true
          description: RFC3339 date where applicable.
    QuranOverview:
      type: object
      required: [module, safety_status, review_status, message, rag, entries]
      properties:
        module:
          type: string
          enum: [quran]
        safety_status:
          $ref: "#/components/schemas/ModuleSafetyStatus"
        review_status:
          type: string
          enum: [unverified, verified, scholar_review_required, rejected, disabled]
        message:
          type: string
        rag:
          $ref: "#/components/schemas/RagReadiness"
        entries:
          type: array
          items:
            $ref: "#/components/schemas/SourceProvenance"
    PrayerOverview:
      type: object
      required: [module, safety_status, review_status, message, rag, windows]
      properties:
        module:
          type: string
          enum: [prayer]
        safety_status:
          $ref: "#/components/schemas/ModuleSafetyStatus"
        review_status:
          type: string
          enum: [unverified, verified, scholar_review_required, rejected, disabled]
        message:
          type: string
        rag:
          $ref: "#/components/schemas/RagReadiness"
        windows:
          type: array
          items:
            $ref: "#/components/schemas/SourceProvenance"
    KnowledgeOverview:
      type: object
      required: [module, safety_status, review_status, message, rag, topics]
      properties:
        module:
          type: string
          enum: [knowledge]
        safety_status:
          $ref: "#/components/schemas/ModuleSafetyStatus"
        review_status:
          type: string
          enum: [unverified, verified, scholar_review_required, rejected, disabled]
        message:
          type: string
        rag:
          $ref: "#/components/schemas/RagReadiness"
        topics:
          type: array
          items:
            $ref: "#/components/schemas/SourceProvenance"
    CommunityOverview:
      type: object
      required: [module, safety_status, review_status, message, rag, channels]
      properties:
        module:
          type: string
          enum: [community]
        safety_status:
          $ref: "#/components/schemas/ModuleSafetyStatus"
        review_status:
          type: string
          enum: [unverified, verified, scholar_review_required, rejected, disabled]
        message:
          type: string
        rag:
          $ref: "#/components/schemas/RagReadiness"
        channels:
          type: array
          items:
            $ref: "#/components/schemas/SourceProvenance"
    RagReadiness:
      type: object
      required: [rag_enabled, rag_index_ready, verified_source_count, review_status]
      properties:
        rag_enabled:
          type: boolean
        rag_index_ready:
          type: boolean
        verified_source_count:
          type: integer
        last_indexed_at:
          type: string
          nullable: true
        review_status:
          type: string
          enum: [unverified, verified, scholar_review_required, rejected, disabled]
    RagStatusResponse:
      type: object
      required: [status, rag_enabled, index_ready, verified_source_count]
      properties:
        status:
          type: string
        rag_enabled:
          type: boolean
        index_ready:
          type: boolean
        verified_source_count:
          type: integer
        last_indexed_at:
          type: string
          nullable: true
    RagSourceItem:
      type: object
      required:
        - module
        - source_type
        - source_name
        - source_reference
        - citation
        - language
        - review_status
        - content_hash
      properties:
        module:
          type: string
        source_type:
          type: string
        source_name:
          type: string
        source_reference:
          type: string
        citation:
          type: string
        language:
          type: string
        review_status:
          type: string
          enum: [unverified, verified, scholar_review_required, rejected, disabled]
        effective_date:
          type: string
          nullable: true
        content_hash:
          type: string
    RagSourcesResponse:
      type: object
      required: [sources]
      properties:
        sources:
          type: array
          items:
            $ref: "#/components/schemas/RagSourceItem"
    RagSearchItem:
      allOf:
        - $ref: "#/components/schemas/RagSourceItem"
        - type: object
          required: [retrieved_at]
          properties:
            retrieved_at:
              type: string
    RagSearchResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RagSearchItem"
    DecisionRequest:
      type: object
      required: [question, selected_module, language, user_subscription_tier]
      properties:
        question:
          type: string
        selected_module:
          type: string
          enum: [quran, prayer, knowledge, community, general_chat, unsupported]
        language:
          type: string
        user_subscription_tier:
          type: string
        safety_context:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
          nullable: true
    DecisionSource:
      allOf:
        - $ref: "#/components/schemas/RagSearchItem"
        - type: object
          required: [similarity_score]
          properties:
            similarity_score:
              type: number
              format: float
    DecisionResponse:
      type: object
      required:
        - answer
        - module
        - language
        - confidence
        - safety_status
        - review_status
        - sources
        - citations
        - generated_from_verified_sources
        - requires_scholar_review
      properties:
        answer:
          type: string
        module:
          type: string
          enum: [quran, prayer, knowledge, community, general_chat, unsupported]
        language:
          type: string
        confidence:
          type: number
          format: float
        safety_status:
          type: string
          enum:
            - safe
            - religious_sensitive
            - medical_sensitive
            - legal_sensitive
            - crisis_sensitive
            - child_sensitive
            - unknown
        review_status:
          type: string
          enum: [unverified, verified, scholar_review_required, rejected, disabled]
        sources:
          type: array
          items:
            $ref: "#/components/schemas/DecisionSource"
        citations:
          type: array
          items:
            type: string
        generated_from_verified_sources:
          type: boolean
        requires_scholar_review:
          type: boolean
    ImanJourneyProgress:
      type: object
      required: [prayer, quran, dhikr]
      properties:
        prayer:
          type: integer
        quran:
          type: integer
        dhikr:
          type: integer
    ImanJourneyFamilyReminder:
      type: object
      required: [consent_granted, notify_family]
      properties:
        consent_granted:
          type: boolean
        reminder_text:
          type: string
          nullable: true
        notify_family:
          type: boolean
    ImanJourneyEvidenceItem:
      type: object
      required: [source_type, source_reference, citation]
      properties:
        source_type:
          type: string
        source_reference:
          type: string
        citation:
          type: string
    ImanJourneySafeFallback:
      type: object
      required: [reason, message]
      properties:
        reason:
          type: string
        message:
          type: string
    ImanJourneyReligiousReminder:
      type: object
      required: [confidence_score, evidence_bundle]
      properties:
        text:
          type: string
          nullable: true
        confidence_score:
          type: number
          format: double
        evidence_bundle:
          type: array
          items:
            $ref: "#/components/schemas/ImanJourneyEvidenceItem"
        safe_fallback:
          allOf:
            - $ref: "#/components/schemas/ImanJourneySafeFallback"
          nullable: true
    ImanJourneyPrivacySettings:
      type: object
      required: [personalization_enabled, reminders_enabled, store_journey_enabled]
      properties:
        personalization_enabled:
          type: boolean
        reminders_enabled:
          type: boolean
        store_journey_enabled:
          type: boolean
    DuaItem:
      type: object
      required: [id, dua_text, is_answered]
      properties:
        id:
          type: string
        dua_text:
          type: string
        is_answered:
          type: boolean
    DuaListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DuaItem"
    UpsertImanJourneyRequest:
      type: object
      required: [today_focus, progress]
      properties:
        journey_date:
          type: string
          format: date
          nullable: true
        today_focus:
          type: string
        continue_yesterday_topic:
          type: string
          nullable: true
        progress:
          $ref: "#/components/schemas/ImanJourneyProgress"
        ask_sakina_today_context:
          type: string
          nullable: true
        tomorrow_follow_up:
          type: string
          nullable: true
        family_reminder:
          $ref: "#/components/schemas/ImanJourneyFamilyReminder"
        religious_reminder_text:
          type: string
          nullable: true
        religious_confidence_score:
          type: number
          format: double
          nullable: true
        evidence_bundle:
          type: array
          items:
            $ref: "#/components/schemas/ImanJourneyEvidenceItem"
    ImanJourneyResponse:
      type: object
      required:
        - journey_date
        - today_focus
        - progress
        - personal_dua_list
        - family_reminder
        - privacy_settings
        - religious_reminder
      properties:
        journey_date:
          type: string
          format: date
        today_focus:
          type: string
        continue_yesterday_topic:
          type: string
          nullable: true
        progress:
          $ref: "#/components/schemas/ImanJourneyProgress"
        personal_dua_list:
          type: array
          items:
            $ref: "#/components/schemas/DuaItem"
        family_reminder:
          $ref: "#/components/schemas/ImanJourneyFamilyReminder"
        ask_sakina_today_context:
          type: string
          nullable: true
        tomorrow_follow_up:
          type: string
          nullable: true
        privacy_settings:
          $ref: "#/components/schemas/ImanJourneyPrivacySettings"
        religious_reminder:
          $ref: "#/components/schemas/ImanJourneyReligiousReminder"
    AddDuaItemRequest:
      type: object
      required: [dua_text]
      properties:
        dua_text:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
  responses:
    FeatureDisabledError:
      description: Module disabled by feature flag
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    SubscriptionRequiredError:
      description: Module requires paid entitlement
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequestError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnauthorizedError:
      description: User scope mismatch or missing scope header
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"